feat: Update mock api
